On my way to conditional solutions
It has been a long-term goal (2 years!) to eliminate the need for two seperate files, one for exercises and one for solutions. The goal is to have a single file where solutions can be toggled by a single flag at the top of the document. This has now been reallized for the first exercises up to and including all complex number exercies.